refactor(secrets): rename HostProvider -> RuntimeSecretsProvider
Name the write-back seam for what the runtime NEEDS, not the transport that
serves it (hostd today; volume/secret-store tomorrow). Rename the interface
HostProvider -> RuntimeSecretsProvider and the class HostdHostProvider -> the
createHostdSecretsProvider factory (matching the repo's functional DI
convention); move host-provider.ts -> secrets-provider.ts. The channel manager
builds the renamed provider. Same secrets-patch RPC on the wire, no behavior
change.
